[refactor] Firehose컴포넌트 모킹을 위한 더미데이타 파일명 수정
