+++ frontend/__hypercloud_mocks__/firehoseMocks.ts
